All: Replaced uint8/uint16 with int; debugging
Decided that using unsigned integers for parameter count, index, and value was causing too many obscure problems to be worth it. Went through and made sure exposed parameters are being properly initialized for each plugin.

+++ Matrix-6G/Matrix-6G 2026/Project_Source_Files/D_TREE_Exposed_Info_P.h
@@ -12,11 +12,11 @@ class Tree_Exposed_Info_P :
 //==============================================================================
 public: Tree_Exposed_Info_P();
 
-public: const uint8 number_for(uint8 param_index) const;
-public: const uint8 byte_index_for(uint8 param_index) const;
-public: const Ctrl_Type ctrl_type_for(const uint8 param_index) const;
-public: const Range_Type range_type_for(const uint8 param_index) const;
-public: const Slider_Display_Type slider_display_type_for(const uint8 param_index) const;
+public: const int number_for(int param_index) const;
+public: const int byte_index_for(int param_index) const;
+public: const Ctrl_Type ctrl_type_for(const int param_index) const;
+public: const Range_Type range_type_for(const int param_index) const;
+public: const Slider_Display_Type slider_display_type_for(const int param_index) const;
 
 //==============================================================================
 private: JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(Tree_Exposed_Info_P)
