+++ P-600-G/P-600-GUI 2026/Project_Source_Files/c_EXP_p.h
@@ -5,7 +5,7 @@
 namespace EXP
 {
 
-	constexpr uint8 choice_count_osc_pitch{ 49 };
-	constexpr uint8 exp_param_count{ 37 };
+	constexpr int choice_count_osc_pitch{ 49 };
+	constexpr int exp_param_count{ 37 };
 
 }
